Fix: execution zctl only from install path
If zctl is installed in a location other than the desktop, it becomes
problematic to run it. `cd`-ing every time - it is not quite usable way.
In fact, you shouldn't bind to the current working folder.

This commit allows you to run the script through PATH just like regular
installed programs. You can symlink to `/usr/share/local` or
`~/.local/bin/` if you like, and everything should work.